Hackathon/ai/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
onnx
onnxruntime
Pillow
numpy
